JDF-450 - Added an interface for Stacks to hide public setters.
